requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
moviepy>=2.0.0
gTTS>=2.3.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pydub>=0.25.1
edge-tts>=6.1.0
schedule>=1.2.0
Flask>=2.3.0
psutil>=5.9.0
